Refactor catalog and recovery code for improved readability and performance
- Simplified snapshot loading logic in `catalog_impl.rs` by combining conditions.
- Replaced `while let` with `for` loop for better clarity in row scanning.
- Enhanced `SystemCatalogSchemas` with a default implementation.
- Streamlined tuple retrieval in `log_recovery.rs` by removing unnecessary references.
- Updated alignment check in `direct_io.rs` to use `is_multiple_of` for clarity.
- Improved task stealing logic in `scheduler.rs` by consolidating conditionals.
- Refined metadata footprint handling in `table_page.rs` to use `?` operator.
- Adjusted error handling in `record.rs` to use `io::Error::other`.
- Enhanced tuple insertion checks in `table_heap.rs` for better readability.

+++ src/common/rid.rs
@@ -246,7 +246,7 @@ impl RID {
         if self.page_id > u32::MAX as PageId {
             return None;
         }
-        let packed = ((self.page_id as u64) << 32) | (self.slot_num as u64);
+        let packed = (self.page_id << 32) | (self.slot_num as u64);
         Some(packed as i64)
     }
